+++ lib/src/real_functions/invertible_real_function.dart
@@ -1,22 +1,33 @@
+import 'package:equatable/equatable.dart';
+
 import '../identity.dart';
 import '../invertible_function.dart';
 
+import 'addition.dart';
+import 'division.dart';
+import 'exponential.dart';
+import 'logarithm.dart';
+import 'multiplication.dart';
+import 'power.dart';
+import 'subtraction.dart';
+
 /// Superclass of all invertible real functions
-abstract class InvertibleRealFunction extends InvertibleFunction<num, num> {
+abstract class InvertibleRealFunction extends InvertibleFunction<num, num>
+    with EquatableMixin {
   /// Constructs a real function
-  InvertibleRealFunction([this._symbol, this._variables]);
+  const InvertibleRealFunction([this._symbol]);
 
   /// Converts any invertible function whose domain ⊆ [num] and codomain ⊆ [num]
   /// into an instance of `InvertibleRealFunction`
   factory InvertibleRealFunction.of(InvertibleFunction<num, num> irf) =>
-      _InvertibleRealFunctionImpl(irf);
+      irf is InvertibleRealFunction ? irf : _InvertibleRealFunctionImpl(irf);
 
   /// Parse [source] as an instance of `InvertibleRealFunction` and return
   factory InvertibleRealFunction.parse(String source) {
     source = source.trim();
     final List<String> parts = source.split(';');
 
-    InvertibleFunction<num, num> irf = Identity<num>();
+    InvertibleRealFunction irf = Identity<num>().toIRF();
 
     for (String part in parts) {
       part = part.trim();
@@ -26,35 +37,29 @@ abstract class InvertibleRealFunction extends InvertibleFunction<num, num> {
       if (f.isEmpty) continue;
 
       final IRFSymbol<dynamic> symbol = IRFSymbol._of(f[0]);
-      final List<String> variables = f.sublist(1);
+      final List<String> props = f.sublist(1);
 
-      if (symbol == null) {
-        throw ArgumentError.value(source, 'source', 'Invalid source');
-      }
-
-      irf >>= symbol.createFunction(variables);
+      irf >>= symbol.createFunction(props);
     }
 
     return InvertibleRealFunction.of(irf);
   }
 
-  final IRFSymbol<dynamic> _symbol;
+  final IRFSymbol<dynamic>? _symbol;
 
-  final List<dynamic> _variables;
+  /// Shorthand for [after]
+  InvertibleRealFunction operator <<(InvertibleFunction<num, num> g) =>
+      InvertibleRealFunction.of(after(g));
 
-  @override
-  InvertibleFunction<W, num> of<W>(InvertibleFunction<W, num> g) {
-    if (g is InvertibleFunction<num, num>) {
-      final InvertibleFunction<num, num> composed =
-          super.of(g as InvertibleFunction<num, num>);
-      return InvertibleRealFunction.of(composed) as InvertibleFunction<W, num>;
-    } else {
-      return super.of(g);
-    }
-  }
+  /// Shorthand for [before]
+  InvertibleRealFunction operator >>(InvertibleFunction<num, num> g) =>
+      InvertibleRealFunction.of(before(g));
 
   @override
-  String toString() => '${_symbol._tokens[0]} ${_variables.join(' ')}';
+  InvertibleRealFunction inverse();
+
+  @override
+  String toString() => '${_symbol?._tokens[0] ?? ''} ${props.join(' ')}'.trim();
 }
 
 class _InvertibleRealFunctionImpl extends InvertibleRealFunction {
@@ -63,17 +68,36 @@ class _InvertibleRealFunctionImpl extends InvertibleRealFunction {
   final InvertibleFunction<num, num> irf;
 
   @override
-  num call(num x) {
-    super.call(x);
-    return irf(x);
-  }
+  List<Object> get props => <Object>[irf];
 
   @override
-  InvertibleRealFunction inverse() =>
+  num valueAt(num x) => irf(x);
+
+  @override
+  _InvertibleRealFunctionImpl inverse() =>
       _InvertibleRealFunctionImpl(irf.inverse());
 
   @override
   String toString() => '$irf';
+}
+
+///
+extension ToIRF on InvertibleFunction<num, num> {
+  /// Converts into an instance of [InvertibleRealFunction]
+  InvertibleRealFunction toIRF() => InvertibleRealFunction.of(this);
+}
+
+///
+extension ComposeListOfIRF on List<InvertibleFunction<num, num>> {
+  /// Returns the composite [InvertibleRealFunction] of
+  /// this [List] of [InvertibleFunction<num, num>]
+  InvertibleRealFunction compose() {
+    InvertibleRealFunction irf = Identity<num>().toIRF();
+    for (final InvertibleFunction<num, num> f in this) {
+      irf >>= f.toIRF();
+    }
+    return irf;
+  }
 }
 
 /// Represents operator(s) of [F]
@@ -85,28 +109,32 @@ abstract class IRFSymbol<F extends InvertibleRealFunction> {
   /// Constructs a symbol for [F]
   ///
   /// [_tokens] is [List] of possible operators
-  IRFSymbol(this._tokens) {
-    final Iterable<String> invalidTokens =
-        _tokens.where((String t) => t.contains(' '));
-    if (invalidTokens.isNotEmpty) {
-      throw ArgumentError.value(
-          _tokens, '_tokens', 'Invalid tokens $invalidTokens');
-    }
-
-    _all.add(this);
-  }
+  const IRFSymbol(this._tokens);
 
   final List<String> _tokens;
 
   /// Creates an instance of [F]
-  F createFunction(List<String> variables);
+  F createFunction(List<String> props);
 
   @override
   String toString() => '$_tokens';
 
-  static final List<IRFSymbol<dynamic>> _all = <IRFSymbol<dynamic>>[];
+  static final List<IRFSymbol<dynamic>> _all = <IRFSymbol<dynamic>>[
+    Plus(),
+    Minus(),
+    Multiply(),
+    Divide(),
+    Pow(),
+    Exp(),
+    Log(),
+  ];
+
+  /// Register a `IRFSymbol`
+  ///
+  /// Registering makes it parsable with [InvertibleRealFunction.parse]
+  static void register(IRFSymbol<dynamic> symbol) => _all.add(symbol);
 
   static IRFSymbol<dynamic> _of(String token) => _all.singleWhere(
-      (IRFSymbol<dynamic> symbol) => symbol._tokens.contains(token),
-      orElse: () => null);
+        (IRFSymbol<dynamic> symbol) => symbol._tokens.contains(token),
+      );
 }